rasp_pi_prj/requirements.txt
#opencv-python
python-dotenv
numpy
fastapi
pydantic
uvicorn
RPi.GPIO